+++ app/folder/[slug]/page.module.scss
