+++ app/folder/[slug]/page.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import dynamic from "next/dynamic";
 
-import styles from "./page.module.css";
+import styles from "./page.module.scss";
 
 const FolderContents = dynamic(
   () => import("components/FolderContents/FolderContents"),
