+++ app/folder/page.module.scss
