+++ app/page.tsx
@@ -4,7 +4,7 @@ import Footer from "@/components/Footer/Footer";
 import Gnb from "@/components/Gnb/Gnb";
 import getUserData from "lib/getUserData";
 
-import styles from "./page.module.css";
+import styles from "./page.module.scss";
 
 export default async function Home() {
   const user = await getUserData();
