+++ app/shared/page.module.scss
@@ -1,4 +1,5 @@
-/* common layouts */
+@use "styles/abstract/variables_color" as var-color;
+
 .inner {
   margin: 0 auto;
   width: 100%;
@@ -16,7 +17,6 @@
   }
 }
 
-/* main */
 .main {
   width: 100%;
   height: 100%;
@@ -80,6 +80,6 @@
 .contents {
   width: 100%;
   height: auto;
-  background: var(--linkbrary-white);
+  background: var-color.$linkbrary-white;
   margin: 0 auto;
 }
