+++ components/AddLink/AddLink.module.scss
@@ -1,3 +1,5 @@
+@use "styles/abstract/variables_color" as var-color;
+
 .formWrapper {
   width: 100%;
   max-width: 800px;
@@ -6,8 +8,8 @@
 .form {
   width: 100%;
   height: 69px;
-  background: #ffffff;
-  border: 1px solid var(--linkbrary-primary);
+  background: var-color.$linkbrary-white;
+  border: 1px solid var-color.$linkbrary-primary;
   border-radius: 15px;
   display: flex;
   justify-content: space-between;
@@ -39,7 +41,7 @@
   font-size: 16px;
   line-height: 24px;
   text-align: left;
-  color: #9fa6b2;
+  color: var-color.$linkbrary-gray-60;
 }
 .input {
   margin: 0;
@@ -61,7 +63,11 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  background: linear-gradient(90.99deg, #6d6afe 0.12%, #6ae3fe 101.84%);
+  background: linear-gradient(
+    90.99deg,
+    var-color.$linkbrary-primary 0.12%,
+    #6ae3fe 101.84%
+  );
   border-radius: 8px;
   color: #f5f5f5;
   font-weight: 600;
