+++ components/AddModal/FolderItem.module.scss
@@ -1,3 +1,5 @@
+@use "styles/abstract/variables_color" as var-color;
+
 .itemWrapper.false {
   height: 40px;
   display: flex;
@@ -10,7 +12,7 @@
   display: flex;
   gap: 8px;
   cursor: pointer;
-  background: #f0f6ff;
+  background: var-color.$linkbrary-bg;
   border-radius: 8px;
   padding: 8px;
   position: relative;
@@ -33,7 +35,7 @@
   line-height: 24px;
   display: flex;
   align-items: center;
-  color: #373740;
+  color: var-color.$linkbrary-gray-100;
 }
 .linkName.true {
   font-weight: 400;
@@ -41,7 +43,7 @@
   line-height: 24px;
   display: flex;
   align-items: center;
-  color: #6d6afe;
+  color: var-color.$linkbrary-primary;
 }
 .linkLength.false {
   font-weight: 400;
@@ -49,7 +51,7 @@
   line-height: 17px;
   display: flex;
   align-items: center;
-  color: #9fa6b2;
+  color: var-color.$linkbrary-gray-60;
 }
 .linkLength.true {
   font-weight: 400;
@@ -57,5 +59,5 @@
   line-height: 17px;
   display: flex;
   align-items: center;
-  color: #9fa6b2;
+  color: var-color.$linkbrary-gray-60;
 }
