+++ components/Card/Card.module.scss
@@ -1,5 +1,7 @@
+@use "styles/abstract/variables_color" as var-color;
+
 .card {
-  background: #ffffff;
+  background: var-color.$linkbrary-white;
   width: 340px;
   height: 335px;
   filter: drop-shadow(0px 5px 25px rgba(0, 0, 0, 0.08));
@@ -73,7 +75,7 @@
   display: flex;
   flex-direction: column;
   align-items: flex-start;
-  background: #ffffff;
+  background: var-color.$linkbrary-white;
   box-shadow: 0px 2px 8px rgba(51, 50, 54, 0.1);
 }
 .cardCaption .info .kebabMenu .popOverWrapper .deleteButton {
