+++ components/DeleteFolderModal/DeleteFolderModal.module.scss
@@ -1,3 +1,5 @@
+@use "styles/abstract/variables_color" as var-color;
+
 .overlay {
   width: 100vw;
   height: 100vh;
@@ -11,7 +13,7 @@
 .modalWrapper {
   width: 360px;
   height: 193px;
-  background-color: #ffffff;
+  background-color: var-color.$linkbrary-white;
   padding: 32px 40px;
   position: relative;
   border: 1px solid #dee2e6;
@@ -25,7 +27,7 @@
   height: 24px;
   cursor: pointer;
 }
-.closeWrapper .image {
+.image {
   width: 100%;
   display: block;
   object-fit: cover;
@@ -43,22 +45,18 @@
   font-size: 20px;
   line-height: 24px;
   text-align: center;
-  color: #373740;
+  color: var-color.$linkbrary-gray-100;
 }
 .subTitle {
   padding-top: 8px;
-  padding-bottom: 40px;
+  padding-bottom: 24px;
   display: block;
   width: 280px;
   font-weight: 400;
   font-size: 14px;
   line-height: 22px;
   text-align: center;
-  color: #9fa6b2;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  word-break: break-all;
+  color: var-color.$linkbrary-gray-60;
 }
 .deleteButton {
   display: flex;
@@ -68,7 +66,7 @@
   padding: 16px 20px;
   width: 280px;
   height: 51px;
-  background: #ff5b56;
+  background: var-color.$linkbrary-red;
   border-radius: 8px;
   font-weight: 600;
   font-size: 16px;
