+++ components/DeleteLinkModal/DeleteLinkModal.module.scss
@@ -1,3 +1,5 @@
+@use "styles/abstract/variables_color" as var-color;
+
 .overlay {
   width: 100vw;
   height: 100vh;
@@ -10,8 +12,8 @@
 }
 .modalWrapper {
   width: 360px;
-  height: 389px;
-  background-color: #ffffff;
+  height: 193px;
+  background-color: var-color.$linkbrary-white;
   padding: 32px 40px;
   position: relative;
   border: 1px solid #dee2e6;
@@ -43,29 +45,24 @@
   font-size: 20px;
   line-height: 24px;
   text-align: center;
-  color: #373740;
+  color: var-color.$linkbrary-gray-100;
 }
 .subTitle {
   padding-top: 8px;
+  padding-bottom: 40px;
   display: block;
   width: 280px;
   font-weight: 400;
   font-size: 14px;
   line-height: 22px;
   text-align: center;
-  color: #9fa6b2;
+  color: var-color.$linkbrary-gray-60;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  word-break: break-all;
 }
-.folderWrapper {
-  padding-top: 32px;
-  padding-bottom: 24px;
-  width: 280px;
-  height: 228px;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  justify-content: center;
-}
-.addButton {
+.deleteButton {
   display: flex;
   flex-direction: row;
   justify-content: center;
@@ -73,7 +70,7 @@
   padding: 16px 20px;
   width: 280px;
   height: 51px;
-  background: linear-gradient(90.99deg, #6d6afe 0.12%, #6ae3fe 101.84%);
+  background: var-color.$linkbrary-red;
   border-radius: 8px;
   font-weight: 600;
   font-size: 16px;
