+++ components/EditFolderModal/EditFolderModal.module.scss
@@ -1,0 +1,92 @@
+@use "styles/abstract/variables_color" as var-color;
+
+.overlay {
+  width: 100vw;
+  height: 100vh;
+  background: rgba(0, 0, 0, 0.4);
+  box-shadow: 0px 4px 25px rgba(0, 0, 0, 0.08);
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+.modalWrapper {
+  width: 360px;
+  height: 244px;
+  background-color: var-color.$linkbrary-white;
+  padding: 35px 40px;
+  position: relative;
+  border: 1px solid var-color.$linkbrary-gray-20;
+  border-radius: 15px;
+}
+.closeWrapper {
+  position: absolute;
+  right: 16px;
+  top: 16px;
+  width: 24px;
+  height: 24px;
+  cursor: pointer;
+}
+.image {
+  width: 100%;
+  display: block;
+  object-fit: cover;
+}
+.contents {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+.title {
+  display: block;
+  width: 280px;
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 24px;
+  text-align: center;
+  color: var-color.$linkbrary-gray-100;
+  padding-bottom: 24px;
+}
+.input {
+  display: block;
+  width: 280px;
+  height: 60px;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  padding: 18px 15px;
+  background: var-color.$linkbrary-white;
+  border: 1px solid var-color.$linkbrary-gray-20;
+  border-radius: 8px;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 24px;
+  color: var-color.$linkbrary-gray-60;
+}
+.buttonWrapper {
+  padding-top: 15px;
+}
+.button {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  align-items: center;
+  padding: 16px 20px;
+  width: 280px;
+  height: 51px;
+  background: linear-gradient(
+    90.99deg,
+    var-color.$linkbrary-primary 0.12%,
+    #6ae3fe 101.84%
+  );
+  border-radius: 8px;
+  font-weight: 600;
+  font-size: 16px;
+  line-height: 19px;
+  display: flex;
+  align-items: center;
+  color: #f5f5f5;
+  cursor: pointer;
+}
