+++ components/FolderList/FolderListItem.module.scss
@@ -1,3 +1,5 @@
+@use "styles/abstract/variables_color" as var-color;
+
 .tabWrapper {
   width: max-content;
   height: 35px;
@@ -5,8 +7,8 @@
   flex-direction: column;
   align-items: center;
   padding: 8px 12px;
-  background: #ffffff;
-  border: 1px solid #6d6afe;
+  background: var-color.$linkbrary-white;
+  border: 1px solid var-color.$linkbrary-primary;
   border-radius: 5px;
   font-weight: 400;
   font-size: 16px;
@@ -14,6 +16,6 @@
   color: #000000;
 }
 .tabWrapper.true {
-  background: #6d6afe;
-  color: #ffffff;
+  background: var-color.$linkbrary-primary;
+  color: var-color.$linkbrary-white;
 }
