+++ components/Footer/Footer.module.scss
@@ -1,4 +1,5 @@
-/* footer layout */
+@use "styles/abstract/variables_color" as var-color;
+
 .inner {
   margin: 0 auto;
   width: 100%;
@@ -19,7 +20,7 @@
 .footer {
   width: 100%;
   height: 16rem;
-  background: #111322;
+  background: var-color.$linkbrary-black;
 }
 .footer .inner {
   height: 1.8rem;
