+++ components/Footer/Footer.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image";
 import Link from "next/link";
 
-import styles from "./Footer.module.css";
+import styles from "./Footer.module.scss";
 
 const Footer = () => {
   return (
