+++ components/Gnb/Gnb.module.scss
@@ -1,4 +1,5 @@
-/* common layouts */
+@use "styles/abstract/variables_color" as var-color;
+
 .inner {
   margin: 0 auto;
   width: 100%;
@@ -16,9 +17,8 @@
   }
 }
 
-/* header */
 .header {
-  background-color: var(--linkbrary-bg-50);
+  background-color: var-color.$linkbrary-bg-50;
   height: 9.3rem;
   position: sticky;
   top: 0;
@@ -40,7 +40,6 @@
   }
 }
 
-/* nav */
 .nav {
   display: flex;
   justify-content: space-between;
@@ -91,7 +90,7 @@
   height: 53px;
   background: linear-gradient(
     90.99deg,
-    var(--linkbrary-primary) 0.12%,
+    var-color.$linkbrary-primary 0.12%,
     #6ae3fe 101.84%
   );
   border-radius: 8px;
@@ -105,7 +104,6 @@
   line-height: 21px;
 }
 @media only screen and (max-width: 1199px) {
-  /* @media (max-width: 863px) { */
   .nav .logo {
     width: 100px;
   }
