--- components/ShareFolderModal/ShareFolderModal.module.css
@@ -1,82 +1,0 @@
-.overlay {
-  width: 100vw;
-  height: 100vh;
-  background: rgba(0, 0, 0, 0.4);
-  box-shadow: 0px 4px 25px rgba(0, 0, 0, 0.08);
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.modalWrapper {
-  width: 360px;
-  height: 209px;
-  background-color: #ffffff;
-  padding: 32px 40px;
-  position: relative;
-  border: 1px solid #dee2e6;
-  border-radius: 15px;
-}
-.closeWrapper {
-  position: absolute;
-  right: 16px;
-  top: 16px;
-  width: 24px;
-  height: 24px;
-  cursor: pointer;
-}
-.contents {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-.title {
-  display: block;
-  width: 280px;
-  font-weight: 700;
-  font-size: 20px;
-  line-height: 24px;
-  text-align: center;
-  color: #373740;
-}
-.subTitle {
-  padding-top: 8px;
-  padding-bottom: 24px;
-  display: block;
-  width: 280px;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 22px;
-  text-align: center;
-  color: #9fa6b2;
-}
-.shareOptions {
-  display: flex;
-  gap: 39px;
-}
-.button {
-  background: #ffffff;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-  cursor: pointer;
-}
-.imageWrapper {
-  width: 42px;
-  height: 42px;
-  position: relative;
-}
-.image {
-  width: 100%;
-  display: block;
-  object-fit: cover;
-}
-.text {
-  font-weight: 400;
-  font-size: 13px;
-  line-height: 15px;
-  text-align: center;
-  color: #373740;
-}
