+++ components/ShareFolderModal/ShareFolderModal.module.scss
@@ -1,3 +1,5 @@
+@use "styles/abstract/variables_color" as var-color;
+
 .overlay {
   width: 100vw;
   height: 100vh;
@@ -10,11 +12,11 @@
 }
 .modalWrapper {
   width: 360px;
-  height: 244px;
+  height: 209px;
   background-color: #ffffff;
-  padding: 35px 40px;
+  padding: 32px 40px;
   position: relative;
-  border: 1px solid #ccd5e3;
+  border: 1px solid #dee2e6;
   border-radius: 15px;
 }
 .closeWrapper {
@@ -24,11 +26,6 @@
   width: 24px;
   height: 24px;
   cursor: pointer;
-}
-.image {
-  width: 100%;
-  display: block;
-  object-fit: cover;
 }
 .contents {
   display: flex;
@@ -43,44 +40,45 @@
   font-size: 20px;
   line-height: 24px;
   text-align: center;
-  color: #373740;
-  padding-bottom: 24px;
+  color: var-color.$linkbrary-gray-100;
 }
-.input {
+.subTitle {
+  padding-top: 8px;
+  padding-bottom: 24px;
   display: block;
   width: 280px;
-  height: 60px;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  padding: 18px 15px;
-  background: #ffffff;
-  border: 1px solid #ccd5e3;
-  border-radius: 8px;
   font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  color: #9fa6b2;
+  font-size: 14px;
+  line-height: 22px;
+  text-align: center;
+  color: var-color.$linkbrary-gray-60;
 }
-.buttonWrapper {
-  padding-top: 15px;
+.shareOptions {
+  display: flex;
+  gap: 39px;
 }
-.addFolderButton {
+.button {
+  background: var-color.$linkbrary-white;
   display: flex;
-  flex-direction: row;
-  justify-content: center;
+  flex-direction: column;
   align-items: center;
-  padding: 16px 20px;
-  width: 280px;
-  height: 51px;
-  background: linear-gradient(90.99deg, #6d6afe 0.12%, #6ae3fe 101.84%);
-  border-radius: 8px;
-  font-weight: 600;
-  font-size: 16px;
-  line-height: 19px;
-  display: flex;
-  align-items: center;
-  color: #f5f5f5;
+  gap: 10px;
   cursor: pointer;
+}
+.imageWrapper {
+  width: 42px;
+  height: 42px;
+  position: relative;
+}
+.image {
+  width: 100%;
+  display: block;
+  object-fit: cover;
+}
+.text {
+  font-weight: 400;
+  font-size: 13px;
+  line-height: 15px;
+  text-align: center;
+  color: var-color.$linkbrary-gray-100;
 }
