+++ package.json
@@ -31,6 +31,7 @@
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.1",
-    "prettier": "^2.8.8"
+    "prettier": "^2.8.8",
+    "sass": "^1.63.1"
   }
 }
