+++ styles/abstract/_variables_color.scss
@@ -1,0 +1,10 @@
+$linkbrary-primary: #6d6afe;
+$linkbrary-red: #ff5b56;
+$linkbrary-black: #111322;
+$linkbrary-white: #ffffff;
+$linkbrary-gray-100: #373740;
+$linkbrary-gray-60: #9fa6b2;
+$linkbrary-gray-20: #ccd5e3;
+$linkbrary-gray-10: #e7effb;
+$linkbrary-bg: #f0f6ff;
+$linkbrary-bg-50: #f0f6ff80;
