+++ styles/base/_reset.scss
@@ -1,0 +1,49 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  outline: none;
+  text-decoration: none;
+  border: none;
+  -webkit-tap-highlight-color: transparent;
+  -webkit-touch-callout: none;
+  -webkit-user-select: none;
+  -khtml-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
+:root {
+  color-scheme: light;
+}
+*:focus {
+  outline: none;
+}
+html {
+  width: 100%;
+  height: 100%;
+  font-size: 62.5%;
+}
+body {
+  width: 100%;
+  font-size: 1.6rem;
+  background-color: $linkbrary-bg;
+}
+a:link,
+a:visited,
+a:hover,
+a:active {
+  text-decoration: none;
+  color: $linkbrary-black;
+}
+#add-portal,
+#delete-link-portal,
+#add-folder-portal,
+#share-folder-portal,
+#edit-folder-portal,
+#delete-folder-portal {
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 9999;
+}
