+++ styles/globals.scss
@@ -1,0 +1,5 @@
+@import "./abstract/mixins";
+@import "./abstract/variables";
+@import "./abstract/variables_color";
+
+@import "./base/reset";
